finish CRUD for all guards without super admin

+++ back-end/backend/app/Http/Controllers/AdminController.php
@@ -13,11 +13,6 @@ use Illuminate\Support\Facades\Storage;
 class AdminController extends Controller
 {
 
-    // function __construct()
-    // {
-    //     $this->middleware('isSuperAdminOwner:admin', ['only' => ['update', 'destroy']]);
-    // }
-
     public function login(Request $request)
     {
 
@@ -30,18 +25,18 @@ class AdminController extends Controller
             if (!Hash::check($request->password, $admin->password)) {
                 return response([
                     'message' => 'Le mot de passe est incorrecte'
-                ], 401);
+                ], 422);
             }
         } else if ($admin = Admin::where('cin', $request->cin_email)->first()) {
             if (!Hash::check($request->password, $admin->password)) {
                 return response([
                     'message' => 'Le mot de passe est incorrect'
-                ], 401);
+                ], 422);
             }
         } else {
             return response([
                 'message' => 'Les identifiants fournis sont incorrects'
-            ], 401);
+            ], 422);
         }
 
         $admin->last_login_date = date('Y-m-d H:i:s');
@@ -61,7 +56,7 @@ class AdminController extends Controller
             'last_name' => 'required|string|max:255',
             'gender' => ['required', Rule::in(['male', 'female'])],
             'email' => 'required|email|unique:admins,email',
-            'cin' => 'required|regex:/^[A-Z]{2}\d+$/|unique:admins,cin',
+            'cin' => 'required|regex:/^[A-Z]{1,2}\d+$/|unique:admins,cin',
             'health_status' => 'nullable|string|max:255',
             'date_of_birth' => 'required|date',
             'blood_type' => ['nullable', Rule::in(['A+', 'A-', 'B+', 'B-', 'AB+', 'AB-', 'O+', 'O-'])],
@@ -78,7 +73,6 @@ class AdminController extends Controller
             $newAdmin->profile_picture = '/picture_profiles/admin/' . $request->cin . '_' . $request->last_name . "-" . $request->first_name . ".jpg";
         }
 
-
         $newAdmin->first_name = $request->first_name;
         $newAdmin->last_name = $request->last_name;
         $newAdmin->gender = $request->gender;
@@ -90,7 +84,7 @@ class AdminController extends Controller
         $newAdmin->blood_type = $request->blood_type;
         $newAdmin->phone_number = $request->phone_number;
         $newAdmin->address = $request->address;
-        $newAdmin->super_admin_id = $request->user()->id;
+        $newAdmin->super_admin_id = $request->user('super_admin')->id;
 
         $newAdmin->save();
 
@@ -105,7 +99,7 @@ class AdminController extends Controller
     public function logout(Request $request)
     {
         try {
-            $request->user()->tokens()->delete();
+            $request->user('admin')->tokens()->delete();
             return response([
                 'message' => 'Déconnexion réussie'
             ], 200);
@@ -121,7 +115,7 @@ class AdminController extends Controller
      */
     public function index(Request $request)
     {
-        return response()->json($request->user()->admins()->latest()->get());
+        return response()->json($request->user('super_admin')->admins()->latest()->get());
     }
 
     /**
@@ -129,10 +123,15 @@ class AdminController extends Controller
      */
     public function show($id)
     {
+        if (request()->user()->cannot('view', Admin::find($id))) {
+            return response()->json([
+                'message' => 'Vous n\'avez pas la permission de voir ce administrateur'
+            ], 401);
+        }
         if (!Admin::find($id)) {
             return response()->json([
                 'message' => 'Admin non trouvé'
-            ]);
+            ], 404);
         }
         return response()->json(Admin::find($id));
     }
@@ -142,10 +141,17 @@ class AdminController extends Controller
      */
     public function update(Request $request, $id)
     {
+
+        if (request()->user()->cannot('update', Admin::find($id))) {
+            return response()->json([
+                'message' => 'Vous n\'avez pas la permission de modifier ce administrateur'
+            ], 401);
+        }
+
         if (!Admin::find($id)) {
             return response()->json([
                 'message' => 'Admin non trouvé'
-            ]);
+            ], 404);
         }
 
         $request->validate([
@@ -198,20 +204,27 @@ class AdminController extends Controller
      */
     public function destroy($id)
     {
+
+        if (request()->user()->cannot('delete', Admin::find($id))) {
+            return response()->json([
+                'message' => 'Vous n\'avez pas la permission de supprimer ce administrateur'
+            ], 401);
+        }
+
         if (!Admin::find($id)) {
             return response()->json([
                 'message' => 'Admin non trouvé'
-            ]);
+            ], 404);
         }
 
         if (!Admin::find($id)->delete()) {
             return response()->json([
                 'message' => 'Admin non trouvé pour la suppression'
-            ]);
+            ], 404);
         }
 
         return response()->json([
-            'message' => 'admin deleted successfuly'
+            'message' => 'Administrateur supprimé avec succès'
         ]);
     }
 }
